Backend features

+++ backend/src/main/java/com/kaleidoscope/backend/auth/repository/EmailVerificationRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface EmailVerificationRepository extends JpaRepository<EmailVerification, Integer> {
+public interface EmailVerificationRepository extends JpaRepository<EmailVerification, Long> {
     Optional<EmailVerification> findByVerificationCode(String verificationCode);
     Optional<EmailVerification> findByEmail(String email);
 }
